requirements.txt
python-pptx==0.6.18
python-docx==0.8.10
click==7.1.2
pydantic[dotenv]
PyMuPDF
pdfplumber
extract_msg
redis
pika==1.2.0